menambah kode peringatan

+++ pertemuan-13/read.php
@@ -47,7 +47,7 @@
       <td><?= $i++ ?></td>
       <td>
         <a href="edit.php?cid=<?= (int)$row['cid']; ?>">Edit</a>
-        <a href="proses_delete.php?cid=<?= (int)$row['cid'];?>">Delete</a>
+        <a onclick="return confirm('Yakin nek ngapus <?= htmlspecialchars($row['cnama']); ?> dengan pesan <?= htmlspecialchars($row['cpesan']); ?>?');" href="proses_delete.php?cid=<?= (int)$row['cid'];?>">Delete</a>
       </td>
       <td><?= $row['cid']; ?></td>
       <td><?= htmlspecialchars($row['cnama']); ?></td>
